+++ task5_2/array.hpp
@@ -1,8 +1,8 @@
 #ifndef ARRAY_HPP
 #define ARRAY_HPP
 
-int *array_generate(int size);
-int array_scalar(int *a, int *b, int size);
-void array_print(int *array, int size);
+double *array_generate(int size);
+double array_scalar(double *a, double *b, int size);
+void array_print(double *array, int size);
 
 #endif
