+++ task5_2/matrix.cpp
@@ -7,9 +7,9 @@
 #include "matrix.hpp"
 
 
-int **matrix_generate(int size)
+double **matrix_generate(int size)
 {
-    int **matrix = new int*[size];
+    double **matrix = new double*[size];
 
     for (int i = 0; i < size; i++) {
         matrix[i] = array_generate(size);
@@ -18,7 +18,7 @@ int **matrix_generate(int size)
 }
 
 
-void matrix_delete(int **matrix, int size)
+void matrix_delete(double **matrix, int size)
 {
     for (int i = 0; i < size; i++) {
         delete[] matrix[i];
@@ -28,7 +28,7 @@ void matrix_delete(int **matrix, int size)
 
 
 #ifdef DEBUG
-void matrix_print(int **matrix, int size)
+void matrix_print(double **matrix, int size)
 {
     for (int i = 0; i < size; i++) {
         array_print(matrix[i], size);
@@ -37,7 +37,7 @@ void matrix_print(int **matrix, int size)
 #endif
 
 
-static void matrix_product(int *result, int **matrix, int *array, int size)
+static void matrix_product(double *result, double **matrix, double *array, int size)
 {
 #pragma omp parallel for
     for (int i = 0; i < size; i++) {
@@ -46,16 +46,16 @@ static void matrix_product(int *result, int **matrix, int *array, int size)
 }
 
 
-static int **matrix_submatrix(int **matrix, int size, int row, int col)
+static double **matrix_submatrix(double **matrix, int size, int row, int col)
 {
-    int **submatrix = new int*[size - 1];
+    double **submatrix = new double*[size - 1];
 
     for (int i = 0, ni = 0; i < size; i++) {
         if (i == row) {
             continue;
         }
 
-        submatrix[ni] = new int[size - 1];
+        submatrix[ni] = new double[size - 1];
         for (int j = 0, nj = 0; j < size; j++) {
             if (j == col) {
                 continue;
@@ -71,7 +71,7 @@ static int **matrix_submatrix(int **matrix, int size, int row, int col)
 }
 
 
-static int matrix_calculate_determinant(int **matrix, int size)
+static double matrix_calculate_determinant(double **matrix, int size)
 {
     if (size == 1) {
         return matrix[0][0];
@@ -80,9 +80,9 @@ static int matrix_calculate_determinant(int **matrix, int size)
         return matrix[0][0] * matrix[1][1] - matrix[0][1] * matrix[1][0];
     }
 
-    int result = 0;
+    double result = 0;
     for (int i = 0, sign = 1; i < size; i++, sign = -sign) {
-        int **submatrix = matrix_submatrix(matrix, size, i, 0);
+        double **submatrix = matrix_submatrix(matrix, size, i, 0);
 
         result += sign * matrix[i][0] * 
             matrix_calculate_determinant(submatrix, size - 1);
@@ -93,17 +93,17 @@ static int matrix_calculate_determinant(int **matrix, int size)
 }
 
 
-static int **matrix_reverse(int **matrix, int size)
+static double **matrix_reverse(double **matrix, int size)
 {
-    int **reverse = new int*[size];
+    double **reverse = new double*[size];
     for (int i = 0; i < size; i++) {
-        reverse[i] = new int[size];
+        reverse[i] = new double[size];
     }
 
-    int determinant = matrix_calculate_determinant(matrix, size);
+    double determinant = matrix_calculate_determinant(matrix, size);
     for (int i = 0; i < size; i++) {
         for (int j = 0; j < size; j++) {
-            int **submatrix = matrix_submatrix(matrix, size, i, j);
+            double **submatrix = matrix_submatrix(matrix, size, i, j);
 
             reverse[j][i] = matrix_calculate_determinant(submatrix, size - 1)
                 / determinant;
@@ -119,10 +119,10 @@ static int **matrix_reverse(int **matrix, int size)
 }
 
 
-int *calculate(int **matrix, int *array, int size)
+double *calculate(double **matrix, double *array, int size)
 {
-    int *result = new int[size];
-    int **reverse = matrix_reverse(matrix, size);
+    double *result = new double[size];
+    double **reverse = matrix_reverse(matrix, size);
 
     matrix_product(result, reverse, array, size);
